pyproject.toml
[tool.poetry]
name = "mpdocs-backend"
version = "0.1.0"
description = "mpdocs backend service"
authors = ["Yakser <iaksanov2044@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
djangorestframework = "^3.14.0"
python-decouple = "^3.8"
django-cors-headers = "^4.3.1"
pre-commit = "^3.6.0"
djangorestframework-simplejwt = "^5.3.1"
django-stubs = {extras = ["compatible-mypy"], version = "^4.2.7"}
djangorestframework-stubs = {extras = ["compatible-mypy"], version = "^3.14.5"}
docxtpl = "^0.16.8"
drf-spectacular = "^0.27.2"
django-rest-polymorphic = "^0.1.10"
drf-spectacular-sidecar = "^2024.5.1"


[tool.poetry.group.dev.dependencies]
django-stubs = {extras = ["compatible-mypy"], version = "^4.2.7"}
djangorestframework-stubs = {extras = ["compatible-mypy"], version = "^3.14.5"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
